-sutvarkyti santykiai su db ir useriu -ui pataisymai -route pridėjimas

+++ app/Http/Controllers/ListingController.php
@@ -43,6 +43,8 @@ class ListingController extends Controller
         $formFields['logo'] = $request->file('logo')->store('logos', 'public');
     }
 
+    $formFields['user_id'] = auth()->id(); 
+
     Listing::create($formFields);
     return redirect('/')->with('message', 'Listing Created');
    }
@@ -53,6 +55,12 @@ class ListingController extends Controller
 
       //update listings data
       public function update(Request $request, Listing $listing){
+
+        //Make sure user is owner
+        if($listing->user_id != auth()->id()){
+            abort(403, 'Unauthorized action');
+        }
+
         $formFields = $request->validate([
             'title' => 'required',
             'company' => ['required'],
@@ -68,12 +76,23 @@ class ListingController extends Controller
         }
     
         $listing->update($formFields);
+
         return back()->with('message', 'Listing updated successfully!');
        }
 
        //Delete listing
        public function destroy(Listing $listing){
+         //Make sure user is owner
+         if($listing->user_id != auth()->id()){
+            abort(403, 'Unauthorized action');
+        }
+
         $listing->delete();
         return redirect('/')->with('message','Listing deleted successfully!');
+       }
+
+       //Manage Listings
+       public function manage(){
+        return view('listings.manage', ['listings' => auth()->user()->listings()->get()]);
        }
 }
